requirements.txt
# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0

# LangChain dependencies
langchain>=0.1.0
langchain-core>=0.1.10
langchain-openai>=0.0.5
langchain-mongodb>=0.0.1

# MongoDB dependencies
pymongo>=4.5.0

# OpenAI API
openai>=1.3.0

# Date/time handling
pytz>=2023.3

# Optional: For development and testing
pytest>=7.4.0
black>=23.7.0
isort>=5.12.0